+++ test/cases/loader-option-preprocessor-default/webpack.config.js
@@ -1,0 +1,39 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    clean: true,
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: {
+          import: './src/index.html',
+          data: {
+            title: 'Home',
+            headline: 'Breaking Bad',
+            name: {
+              firstname: 'Walter',
+              lastname: 'White',
+            },
+          },
+        },
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(html)$/,
+        loader: HtmlBundlerPlugin.loader,
+        // test the Eta/EJS as the default template engine used in the default preprocessor
+      },
+    ],
+  },
+};